se modularizan las funciones
