+++ pages/detail.html
@@ -47,7 +47,7 @@
             </footer>
         </div> 
         <script src="../assets/moivestackData/data.js"></script>
-        <script src="../scripts/details.js"></script>
+        <script type="module" src="../scripts/details.js"></script>
 
     </body>
