+++ pages/movies.html
@@ -71,6 +71,6 @@
         </footer>
     </div>
     <script src="../assets/moivestackData/data.js"></script>
-    <script src="../scripts/movies.js"></script>
+    <script type="module" src="../scripts/movies.js"></script>
 </body>
 </html>
