+++ scripts/index.js
@@ -1,6 +1,6 @@
 const openMenu = document.getElementById("menu");
 const drop = document.getElementById("dropMenu");
-const close = document.getElementById("close");
+const $close = document.getElementById("close");
 
 function dropMenu(){
     drop.classList.toggle("hidden"); 
@@ -8,8 +8,8 @@ function dropMenu(){
     close.classList.toggle("hidden");
 }
 
-openMenu.addEventListener("click",dropMenu );
+openMenu.addEventListener("click",dropMenu);
 
 
 
-close.addEventListener("click", dropMenu)
+$close.addEventListener("click", dropMenu)
